requirements.txt
cycler==0.10.0
fire==0.1.3
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
rawpy==0.13.1
six==1.12.0
